vhc: i32 is the index lane — i64 kernels don't exist on a stock DX12 box
The guest channel pinned IntElem = i64, so every index tensor — arange
position ids, the embedding gather and its select_assign gradient
accumulation, the target gather — rode the compute@2 wire as i64 kernels. The
5090 exposes SHADER_INT64 under Vulkan but not under DX12: wgpu compiles DX12
shaders with FXC by default (SM 5.1), the feature is DXC-only, and cubecl's
own instance construction never reads WGPU_DX12_COMPILER — so the lane refused
select_assign_kernel_f_f32_i_i64 at validation, thirty slices into a round
(backend-lane audit 2026-07-28, D2).

i32 is the honest dtype, not a workaround: every index space the module
addresses fits it — vocab 32,768, seq 2,048, largest per-tensor numel
50,331,648 < 2^31 — and an index dtype changes no f32 value. The trainer
goldens (det digests, theta trajectory) reproduce bit-for-bit, which is the
proof. The plan now declares what the module actually launches (f32/i32), the
guest builds its id rows as i32 (4 bytes each, and the linear-memory term says
so), and the module hashes re-pin through build-guests.

The other half is the disarmed gate the audit called D4: the dev-minted
profiles copied one dtype list across every backend class, so a dx12-class
profile claimed i64, the planner's typed-refusal check at admission could
never fire, and the capability gap surfaced as a mid-round ComputeFault
instead. The fixture the minting authors through now states coverage per
class, gated exactly as cubecl gates the type — on SHADER_INT64, which
Vulkan/Metal/CUDA/CPU have and stock DX12 does not — so a plan needing i64
refuses typed at admission on the one lane that cannot run it. The mirror
channels and the wire-walking regression test extend to the int lane in
lockstep.

+++ crates/vhc/guests/tiny-llama/src/model.rs
@@ -414,7 +414,7 @@ impl<B: AutodiffBackend> TinyLlamaModel<B> {
     /// forward loss *scalar*'s summation order (`rows` terms instead of `rows × vocab` terms, the
     /// extra ones being exact zeros); the scalar is consumed only as `backward()`'s root, which is
     /// seeded independently of its value.
-    fn target_logprobs(&self, logsm: Tensor<B, 2>, tgt: Vec<i64>, rows: usize) -> Tensor<B, 2> {
+    fn target_logprobs(&self, logsm: Tensor<B, 2>, tgt: Vec<i32>, rows: usize) -> Tensor<B, 2> {
         let targets = Tensor::<B, 1, Int>::from_data(TensorData::new(tgt, [rows]), &self.device)
             .reshape([rows, 1]);
         logsm.gather(1, targets)
@@ -438,8 +438,11 @@ impl<B: AutodiffBackend> TinyLlamaModel<B> {
         let mut tgt = Vec::with_capacity(rows);
         for bi in 0..b {
             for si in 0..s {
-                inp.push(i64::from(tokens[bi * seq + si]));
-                tgt.push(i64::from(tokens[bi * seq + si + 1]));
+                // Token ids are vocabulary indices (< 2¹⁵ at any admitted geometry), so i32 is
+                // lossless — and it is the module's index dtype: i64 kernels don't exist on a
+                // stock DX12 lane (SHADER_INT64 is DXC-only there; backend-lane audit D2).
+                inp.push(tokens[bi * seq + si] as i32);
+                tgt.push(tokens[bi * seq + si + 1] as i32);
             }
         }
